requirements.txt
cycler==0.10.0
kiwisolver==1.0.1
matplotlib==2.2.2
numpy==1.14.3
Pillow==5.1.0
pyparsing==2.2.0
PyQt5==5.10.1
python-dateutil==2.7.3
pytz==2018.4
sip==4.19.8
six==1.11.0
